Allow node names to be suffixed with numbers
I had to remove the use of String#rindex and replace it with a custom
function that finds the earliest match that ends at a given index. For
example, take a string "a-b2" and a regex /\w[0-9]?/. String#rindex will
return 3, while the new String#index_of_earliest_match_ending_at will
return 2.