Don't parse graphs in test setup method
This makes it easier to debug a single test, because only the code for
that one test will run. Most graphs were moved down into the single test
method that uses them. Graphs that are used by more than one test were
left in setup, but changed to just be strings instead of parsed graphs.
The tests that use each graph now handle the parsing.